Fixing test infrastructure.
There was a pretty substantive data management issue with the original testing organization. I had to rearrange things quite a bit so that the test objects would be alive for the duration of the tests.
The updates keeps things alive in a way that the testing can complete before destruction occurrs. The new structure places each tester as its own object that runs the tests, but this was not the original hope. The original hope was that there was 1 tester and then groups of test objects passed around - this is not necessarily possible though because of the information required for any given test.